services/requirements.txt
# This requirements file is like a master-requirements file which installs all the dependencies needed for all the lambdas and the dependencies needed for testing.
-r structured_mysql/batch_job_aurora/requirements.txt
-r events_slack_app/slack_response/requirements.txt
-r poller/daily_poller/requirements.txt

# The lambdas don't need boto in their requirements file because it's already installed on amazon lambda.
atomicwrites==1.3.0
attrs==19.1.0
boto3==1.9.178
botocore==1.12.178
docutils==0.14
importlib-metadata==0.18
jmespath==0.9.4
more-itertools==7.1.0
packaging==19.0
pluggy==0.12.0
py==1.8.0
pyparsing==2.4.0
pytest==4.6.3
python-dateutil==2.8.0
s3transfer==0.2.1
six==1.12.0
urllib3==1.25.3
wcwidth==0.1.7
zipp==0.5.1
ruamel.yaml==0.16.5
google-api-python-client==1.7.11
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.0
airtable-python-wrapper==0.12.0